Initialize TargetView
Initialize targetview.
Clear up.
